feat(suite): extract firmware upgrade helpers
