+++ packages/suite/src/components/firmware/index.tsx
@@ -2,10 +2,6 @@ export { ReconnectDevicePrompt } from './ReconnectDevicePrompt';
 export { FirmwareOffer } from './FirmwareOffer';
 export { FirmwareInstallation } from './FirmwareInstallation';
 export { FirmwareInitial } from './FirmwareInitial';
-export { Fingerprint } from './Fingerprint';
-export { FirmwareProgressBar } from './FirmwareProgressBar';
 export { SelectCustomFirmware } from './SelectCustomFirmware';
 export { FirmwareInstallationProgressCheck } from './ProgressCheck/FirmwareInstallationProgressCheck';
 export { RotatingPhrases } from './RotatingPhrases';
-export { FirmwareWarningsList } from './FirmwareWarningsList';
-export { FirmwareWipeWarning } from './FirmwareWipeWarning';
