+++ packages/suite/src/views/firmware/FirmwareCustom.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 
+import { useFirmwareDesktopUpdate } from '@suite/firmware-upgrade';
 import { Translation } from '@suite/intl';
 
 import { SelectCustomFirmware } from 'src/components/firmware';
 import { FirmwareLowBatteryModal } from 'src/components/firmware/FirmwareLowBatteryModal';
-import { useFirmwareDesktopUpdate } from 'src/hooks/suite/useFirmwareDesktopUpdate';
 
 import { FirmwareModal } from './FirmwareModal';
