+++ packages/suite/src/views/firmware/FirmwareModal.tsx
@@ -1,5 +1,9 @@
 import { type ReactNode, useState } from 'react';
 
+import {
+    useFirmwareDesktopUpdate,
+    useFirmwareInstallationProgressCheck,
+} from '@suite/firmware-upgrade';
 import { closeModal } from '@suite/modal';
 import { closeModalApp } from '@suite/router';
 import { selectSelectedDevice } from '@suite-common/device';
@@ -7,16 +11,15 @@ import { acquireDevice } from '@suite-common/wallet-core';
 import { Modal } from '@trezor/components';
 import { exhaustive } from '@trezor/type-utils';
 
+import { FirmwareInstallationProgressCheck } from 'src/components/firmware';
 import { ThpPairingStep } from 'src/components/firmware/ThpPairingStep/ThpPairingStep';
-import { useDispatch, useFirmwareInstallationProgressCheck, useSelector } from 'src/hooks/suite';
-import { useFirmwareDesktopUpdate } from 'src/hooks/suite/useFirmwareDesktopUpdate';
+import { useDispatch, useSelector } from 'src/hooks/suite';
 
 import { StepCheckSeed } from './Steps/StepCheckSeed';
 import { StepDone } from './Steps/StepDone';
 import { StepError } from './Steps/StepError';
 import { StepInitial } from './Steps/StepInitial';
 import { StepStarted } from './Steps/StepStarted';
-import { FirmwareInstallationProgressCheck } from '../../components/firmware';
 
 type FirmwareModalProps = {
     children: ReactNode;
