+++ suite/firmware-upgrade/src/FirmwareWipeWarning.tsx
