+++ suite/firmware-upgrade/src/index.ts
@@ -1,1 +1,13 @@
+export { Fingerprint } from './Fingerprint';
+export { FirmwareProgressBar } from './FirmwareProgressBar';
 export { FirmwareUpgradeNeededModal } from './FirmwareUpgradeNeededModal';
+export { FirmwareWarningsList } from './FirmwareWarningsList';
+export { FirmwareWipeWarning } from './FirmwareWipeWarning';
+export {
+    getFormattedFingerprint,
+    getSuiteFirmwareTypeString,
+    parseFirmwareFormat,
+    validateFirmware,
+} from './firmwareUtils';
+export { useFirmwareDesktopUpdate } from './useFirmwareDesktopUpdate';
+export { useFirmwareInstallationProgressCheck } from './useFirmwareInstallationProgressCheck';
