+++ suite/firmware-upgrade/src/useFirmwareInstallationProgressCheck.ts
